+++ general.py
@@ -7,9 +7,12 @@ class Options:
     female_population_percentage: float
     male_population_count: int
     female_population_count: int
-    max_friends_per_person: int
-    male_friends_percentage: float
-    female_friends_percentage: float
+    max_friends_per_male: int
+    max_friends_per_female: int
+    male_on_male_friends_percentage: float
+    male_on_female_friends_percentage: float
+    female_on_male_friends_percentage: float
+    female_on_female_friends_percentage: float
     sample_size: int
     iterations_count: int
     unbiased_analysis: bool
@@ -30,10 +33,18 @@ def print_log(options: Options):
         )
     )
     print(
-        colored("Friends/Person:\t\t", "black", attrs=["bold"])
-        + colored(f"{options.max_friends_per_person} Max", "blue")
+        colored("Friends/Male:\t\t", "black", attrs=["bold"])
+        + colored(f"{options.max_friends_per_male} Max", "blue")
         + colored(
-            f" ({options.male_friends_percentage * 100}% Male + {options.female_friends_percentage * 100}% Female)",
+            f" ({options.male_on_male_friends_percentage * 100}% Male + {options.male_on_female_friends_percentage * 100}% Female)",
+            attrs=["bold"],
+        )
+    )
+    print(
+        colored("Friends/Female:\t\t", "black", attrs=["bold"])
+        + colored(f"{options.max_friends_per_female} Max", "blue")
+        + colored(
+            f" ({options.female_on_male_friends_percentage * 100}% Male + {options.female_on_female_friends_percentage * 100}% Female)",
             attrs=["bold"],
         )
     )
@@ -72,9 +83,18 @@ def get_options(options) -> Options:
     obj.female_population_count = int(
         options.population * obj.female_population_percentage
     )
-    obj.max_friends_per_person = options.maxFriendsPerPerson
-    obj.male_friends_percentage = round(options.maleFriends * 1000) / 100000
-    obj.female_friends_percentage = round((100 - options.maleFriends) * 1000) / 100000
+    obj.max_friends_per_male = options.maxFriendsPerMale
+    obj.max_friends_per_female = options.maxFriendsPerFemale
+    obj.male_on_male_friends_percentage = round(options.maleMaleFriends * 1000) / 100000
+    obj.male_on_female_friends_percentage = (
+        round((100 - options.maleMaleFriends) * 1000) / 100000
+    )
+    obj.female_on_female_friends_percentage = (
+        round(options.femaleFemaleFriends * 1000) / 100000
+    )
+    obj.female_on_male_friends_percentage = (
+        round((100 - options.femaleFemaleFriends) * 1000) / 100000
+    )
     obj.sample_size = options.sampleSize
     obj.iterations_count = options.iterations
     obj.unbiased_analysis = options.unbiasedAnalysis